backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
bcrypt<4.0.0
python-multipart
email-validator

